refactor: split web3 and services bean declarations in different configuration files
