+++ hedera-mirror-web3/src/main/java/com/hedera/mirror/web3/evm/config/ServicesConfiguration.java
@@ -1,0 +1,268 @@
+/*
+ * Copyright (C) 2019-2023 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hedera.mirror.web3.evm.config;
+
+import com.hedera.mirror.web3.evm.pricing.RatesAndFeesLoader;
+import com.hedera.mirror.web3.evm.properties.MirrorNodeEvmProperties;
+import com.hedera.node.app.service.evm.contracts.execution.EvmProperties;
+import com.hedera.services.contracts.gascalculator.GasCalculatorHederaV22;
+import com.hedera.services.fees.BasicHbarCentExchange;
+import com.hedera.services.fees.FeeCalculator;
+import com.hedera.services.fees.HbarCentExchange;
+import com.hedera.services.fees.calc.OverflowCheckingCalc;
+import com.hedera.services.fees.calculation.BasicFcfsUsagePrices;
+import com.hedera.services.fees.calculation.TxnResourceUsageEstimator;
+import com.hedera.services.fees.calculation.UsageBasedFeeCalculator;
+import com.hedera.services.fees.calculation.UsagePricesProvider;
+import com.hedera.services.fees.calculation.token.txns.TokenAssociateResourceUsage;
+import com.hedera.services.fees.calculation.token.txns.TokenDissociateResourceUsage;
+import com.hedera.services.fees.calculation.utils.AccessorBasedUsages;
+import com.hedera.services.fees.calculation.utils.PricedUsageCalculator;
+import com.hedera.services.fees.pricing.AssetsLoader;
+import com.hedera.services.hapi.fees.usage.EstimatorFactory;
+import com.hedera.services.hapi.fees.usage.TxnUsageEstimator;
+import com.hedera.services.ledger.TransferLogic;
+import com.hedera.services.store.contracts.precompile.Precompile;
+import com.hedera.services.store.contracts.precompile.PrecompileMapper;
+import com.hedera.services.store.contracts.precompile.SyntheticTxnFactory;
+import com.hedera.services.store.contracts.precompile.codec.EncodingFacade;
+import com.hedera.services.store.contracts.precompile.impl.AssociatePrecompile;
+import com.hedera.services.store.contracts.precompile.impl.DissociatePrecompile;
+import com.hedera.services.store.contracts.precompile.impl.MintPrecompile;
+import com.hedera.services.store.contracts.precompile.impl.MultiAssociatePrecompile;
+import com.hedera.services.store.contracts.precompile.impl.MultiDissociatePrecompile;
+import com.hedera.services.store.contracts.precompile.utils.PrecompilePricingUtils;
+import com.hedera.services.txn.token.AssociateLogic;
+import com.hedera.services.txn.token.BurnLogic;
+import com.hedera.services.txn.token.DissociateLogic;
+import com.hedera.services.txn.token.MintLogic;
+import com.hedera.services.txns.crypto.ApproveAllowanceLogic;
+import com.hedera.services.txns.crypto.AutoCreationLogic;
+import com.hedera.services.txns.crypto.DeleteAllowanceLogic;
+import com.hedera.services.txns.validation.ContextOptionValidator;
+import com.hedera.services.txns.validation.OptionValidator;
+import com.hedera.services.utils.accessors.AccessorFactory;
+import com.hederahashgraph.api.proto.java.HederaFunctionality;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Spring configuration for beans related to com.hedera.services components
+ * */
+@Configuration
+public class ServicesConfiguration {
+
+    @Bean
+    GasCalculatorHederaV22 gasCalculatorHederaV22(
+            BasicFcfsUsagePrices usagePricesProvider, BasicHbarCentExchange hbarCentExchange) {
+        return new GasCalculatorHederaV22(usagePricesProvider, hbarCentExchange);
+    }
+
+    @Bean
+    BasicFcfsUsagePrices basicFcfsUsagePrices(RatesAndFeesLoader ratesAndFeesLoader) {
+        return new BasicFcfsUsagePrices(ratesAndFeesLoader);
+    }
+
+    @Bean
+    AccessorBasedUsages accessorBasedUsages() {
+        return new AccessorBasedUsages();
+    }
+
+    @Bean
+    OverflowCheckingCalc overflowCheckingCalc() {
+        return new OverflowCheckingCalc();
+    }
+
+    @Bean
+    PricedUsageCalculator pricedUsageCalculator(
+            AccessorBasedUsages accessorBasedUsages, OverflowCheckingCalc overflowCheckingCalc) {
+        return new PricedUsageCalculator(accessorBasedUsages, overflowCheckingCalc);
+    }
+
+    @Bean
+    EstimatorFactory estimatorFactory() {
+        return TxnUsageEstimator::new;
+    }
+
+    @Bean
+    TokenAssociateResourceUsage tokenAssociateResourceUsage(final EstimatorFactory estimatorFactory) {
+        return new TokenAssociateResourceUsage(estimatorFactory);
+    }
+
+    @Bean
+    TokenDissociateResourceUsage tokenDissociateResourceUsage(final EstimatorFactory estimatorFactory) {
+        return new TokenDissociateResourceUsage(estimatorFactory);
+    }
+
+    @Bean
+    UsageBasedFeeCalculator usageBasedFeeCalculator(
+            HbarCentExchange hbarCentExchange,
+            UsagePricesProvider usagePricesProvider,
+            PricedUsageCalculator pricedUsageCalculator,
+            List<TxnResourceUsageEstimator> txnResourceUsageEstimators) {
+        // queryUsageEstimators and txnResourceUsegaEstimator will be implemented in separate PR
+        final Map<HederaFunctionality, List<TxnResourceUsageEstimator>> txnUsageEstimators = new HashMap<>();
+
+        for (final var estimator : txnResourceUsageEstimators) {
+            if (estimator.toString().contains("TokenAssociate")) {
+                txnUsageEstimators.put(HederaFunctionality.TokenAssociateToAccount, List.of(estimator));
+            }
+            if (estimator.toString().contains("TokenDissociate")) {
+                txnUsageEstimators.put(HederaFunctionality.TokenDissociateFromAccount, List.of(estimator));
+            }
+        }
+
+        return new UsageBasedFeeCalculator(
+                hbarCentExchange,
+                usagePricesProvider,
+                pricedUsageCalculator,
+                Collections.emptySet(),
+                txnUsageEstimators);
+    }
+
+    @Bean
+    AssetsLoader assetsLoader() {
+        return new AssetsLoader();
+    }
+
+    @Bean
+    AccessorFactory accessorFactory() {
+        return new AccessorFactory();
+    }
+
+    @Bean
+    PrecompilePricingUtils precompilePricingUtils(
+            final AssetsLoader assetsLoader,
+            final BasicHbarCentExchange exchange,
+            final FeeCalculator feeCalculator,
+            final BasicFcfsUsagePrices resourceCosts,
+            final AccessorFactory accessorFactory) {
+        return new PrecompilePricingUtils(assetsLoader, exchange, feeCalculator, resourceCosts, accessorFactory);
+    }
+
+    @Bean
+    BasicHbarCentExchange basicHbarCentExchange(RatesAndFeesLoader ratesAndFeesLoader) {
+        return new BasicHbarCentExchange(ratesAndFeesLoader);
+    }
+
+    @Bean
+    AssociatePrecompile associatePrecompile(
+            final PrecompilePricingUtils precompilePricingUtils,
+            final SyntheticTxnFactory syntheticTxnFactory,
+            final AssociateLogic associateLogic) {
+        return new AssociatePrecompile(precompilePricingUtils, syntheticTxnFactory, associateLogic);
+    }
+
+    @Bean
+    MultiAssociatePrecompile multiAssociatePrecompile(
+            final PrecompilePricingUtils precompilePricingUtils,
+            final SyntheticTxnFactory syntheticTxnFactory,
+            final AssociateLogic associateLogic) {
+        return new MultiAssociatePrecompile(precompilePricingUtils, syntheticTxnFactory, associateLogic);
+    }
+
+    @Bean
+    PrecompileMapper precompileMapper(final Set<Precompile> precompiles) {
+        return new PrecompileMapper(precompiles);
+    }
+
+    @Bean
+    OptionValidator optionValidator(final MirrorNodeEvmProperties properties) {
+        return new ContextOptionValidator(properties);
+    }
+
+    @Bean
+    EncodingFacade encodingFacade() {
+        return new EncodingFacade();
+    }
+
+    @Bean
+    SyntheticTxnFactory syntheticTxnFactory() {
+        return new SyntheticTxnFactory();
+    }
+
+    @Bean
+    DeleteAllowanceLogic deleteAllowanceLogic() {
+        return new DeleteAllowanceLogic();
+    }
+
+    @Bean
+    AssociateLogic associateLogic(MirrorNodeEvmProperties mirrorNodeEvmProperties) {
+        return new AssociateLogic(mirrorNodeEvmProperties);
+    }
+
+    @Bean
+    ApproveAllowanceLogic approveAllowanceLogic() {
+        return new ApproveAllowanceLogic();
+    }
+
+    @Bean
+    AutoCreationLogic autocreationLogic(
+            FeeCalculator feeCalculator, EvmProperties evmProperties, SyntheticTxnFactory syntheticTxnFactory) {
+        return new AutoCreationLogic(feeCalculator, evmProperties, syntheticTxnFactory);
+    }
+
+    @Bean
+    MintPrecompile mintPrecompile(
+            PrecompilePricingUtils precompilePricingUtils,
+            EncodingFacade encodingFacade,
+            SyntheticTxnFactory syntheticTxnFactory,
+            MintLogic mintLogic) {
+        return new MintPrecompile(precompilePricingUtils, encodingFacade, syntheticTxnFactory, mintLogic);
+    }
+
+    @Bean
+    MintLogic mintLogic(OptionValidator optionValidator) {
+        return new MintLogic(optionValidator);
+    }
+
+    @Bean
+    BurnLogic burnLogic(OptionValidator optionValidator) {
+        return new BurnLogic(optionValidator);
+    }
+
+    @Bean
+    DissociateLogic dissociateLogic() {
+        return new DissociateLogic();
+    }
+
+    @Bean
+    TransferLogic transferLogic(AutoCreationLogic autoCreationLogic) {
+        return new TransferLogic(autoCreationLogic);
+    }
+
+    @Bean
+    DissociatePrecompile dissociatePrecompile(
+            final PrecompilePricingUtils precompilePricingUtils,
+            final SyntheticTxnFactory syntheticTxnFactory,
+            final DissociateLogic dissociateLogic) {
+        return new DissociatePrecompile(precompilePricingUtils, syntheticTxnFactory, dissociateLogic);
+    }
+
+    @Bean
+    MultiDissociatePrecompile multiDissociatePrecompile(
+            final PrecompilePricingUtils precompilePricingUtils,
+            final SyntheticTxnFactory syntheticTxnFactory,
+            final DissociateLogic dissociateLogic) {
+        return new MultiDissociatePrecompile(precompilePricingUtils, syntheticTxnFactory, dissociateLogic);
+    }
+}
